+++ src/models/CssSnippet.ts
@@ -1,0 +1,6 @@
+export interface CssSnippet {
+  key: string;
+  description: string;
+  type: string;
+  component: string;
+}