+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "vscode-msgraph-essentials" extension will be documented in this file.
 
+## [0.0.3]
+
+- CSS snippets added for quickly changing component styles
+- HTML snippets added for easy insert of the MGT components in HTML
+- Autocompletion added for all the CSS variables which can be used to change the styling of the components
+
 ## [0.0.2]
 
 - Update to the `mgt-react-tmp` snippet